requirements.txt
streamlit==1.30.0
torch==2.2.1
torchvision==0.17.1
Pillow==10.3.0
numpy==1.26.4
matplotlib==3.8.4
opencv-python
scikit-learn
pandas
